minor changes
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/branches/pcl-1.x@5044 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ io/src/openni_camera/openni_image_bayer_grbg.cpp
@@ -263,7 +263,7 @@ void ImageBayerGRBG::fillGrayscale (unsigned width, unsigned height, unsigned ch
       }
     }
     else
-      THROW_OPENNI_EXCEPTION ("Unknown Debayering method: %d", (int)debayering_method_);
+      THROW_OPENNI_EXCEPTION ("Unknown Debayering method: %d", debayering_method_);
 
   }
   else // downsampling
@@ -1395,7 +1395,7 @@ void ImageBayerGRBG::fillRGB (unsigned width, unsigned height, unsigned char* rg
       //rgb_pixel[rgb_line_step + 5] = bayer_pixel[line_step];
     }
     else
-      THROW_OPENNI_EXCEPTION ("Unknwon debayering method: %d", (int)debayering_method_);
+      THROW_OPENNI_EXCEPTION ("Unknwon debayering method: %d", debayering_method_);
   }
   else
   {
